+++ test/compute-engine/collections.test.ts
@@ -458,10 +458,10 @@ describe('Drop 2', () => {
 });
 
 describe('INDEXABLE OPERATIONS', () => {
-  test('At', () =>
+  test('At with positive index', () =>
     expect(evaluate(['At', list, 1])).toMatchInlineSnapshot(`7`));
 
-  test('At', () =>
+  test('At with negative index', () =>
     expect(evaluate(['At', list, -2])).toMatchInlineSnapshot(`Nothing`));
 
   test('At', () =>